Update tsjava options in package.json.
We're now using ts-java as of redseal/feature/ts-java-candidate-set-of-classes-by-listing-jar-files_89437380,
which was a significant change in how ts-java decides which classes to include in the output java.d.ts file.

Note that 'com.tinkerpop.' is no longer a whitelisted package, as that leads to generating
a java.d.ts file that is over 100K lines long, and clearly slows down tsc compiles. Instead,
we explicitly list the tinkerpop classes we want to include. This list may need to be tweaked
as we add new use cases.