Use dts-bundle to create bundle.d.ts.
This is a failed attempt to use this tool to combine the various declaration files into one coherent declaration.  It
fails primarily because it does not preserve the reference to java.d.ts, so all Java references are undefined.  It is
also less than ideal because it inlines bluebird.d.ts, rather than preserve its reference.